% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dbms.R, R/get_dbms_backup.R
\name{get_dbms}
\alias{get_dbms}
\title{Collecting DBMS popularity data}
\usage{
get_dbms(type = c("all", "Content store", "Document store",
"Event Store", "Graph DBMS", "Key-value store", "Multivalue DBMS",
"Native XML DBMS", "Navigational DBMS", "Object oriented DBMS", "RDF store",
"Relational DBMS", "Search engine", "Time Series DBMS", "Wide column store"))

get_dbms(type = c("all", "Content store", "Document store", "Event Store", "Graph DBMS", "Key-value store", "Multivalue DBMS", "Native XML DBMS", "Navigational DBMS", "Object oriented DBMS", "RDF store", "Relational DBMS", "Search engine", "Time Series DBMS", "Wide column store"))
}
\arguments{
\item{type}{of DBMS; default: "all"}

\item{type}{of DBMS; default: "all"}
}
\value{
data frame

data frame
}
\description{
Function \code{get_dbms} scrapes data from db-engines.com

Function \code{get_dbms} scrapes data from db-engines.com
}
\examples{
\dontrun{
Top 10 Relational Database Management Systems
dbms <- get_dbms("Relational DBMS")
head(dbms, 10)
}
\dontrun{
Top 10 Relational Database Management Systems
dbms <- get_dbms("Relational DBMS")
head(dbms, 10)
}
}
\author{
Krzysztof Słomczyński

Krzysztof Słomczyński
}

