% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ids.R
\name{get_ids}
\alias{get_ids}
\title{Geting indexes of offers that match selected criteria}
\usage{
get_ids(offers, patterns = c("data sci", "stat.st", "anality",
  "analiz. danych", "[^A-Za-z]SAS",
  "[^0-9][^0-9A-Za-zóÓ+]R[^0-9A-Za-zóÓęĘśŚćĆ&“”/ö][^&]",
  "big.+data", "data.+mining", "tableau", "SPSS", "stata", "cognos", "azure",
  "machine learning", "spark", "hadoop", "matlab"), case = c(T, T, T, T, T, T,
  F, T, T, T, T, T, T, T, T, T, T, T, T, T, T, T))
}
\arguments{
\item{offers}{vector with offers descriptions}

\item{patterns}{patterns to be found in descriptions}

\item{case}{should the pattern be identified case sensitive?}
}
\value{
data frame
}
\description{
Function \code{get_ids} returns ids of offers that match selected criteria.
}
\examples{
\dontrun{
offers <- get_offers(description=TRUE)
get_ids(offers$description)
}
}
\author{
Przemyslaw Biecek
}

